requirements.txt
Flask==3.0.2
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
python-barcode==0.15.1
Pillow==10.2.0
pandas==2.2.0
openpyxl==3.1.2
python-dotenv==1.0.1
SQLAlchemy==2.0.25
python-dateutil==2.8.2
schedule==1.2.1
gunicorn==20.1.0
pytest==7.3.1
marshmallow==3.19.0
email-validator==2.0.0
werkzeug==2.3.3